Improve timing precision with microsecond resolution
- Add hal::micros() to platform abstraction layer
- Switch effect_processor to use micros() for deltaTime calculation
- Remove artificial frame rate cap from main.cpp loop
- Add test helpers for microsecond time control

This improves animation smoothness by:
1. Reducing timing error from ±12% (millis) to <0.1% (micros) at 120 FPS
2. Removing duplicate timing logic between main.cpp and effect_processor
3. Allowing maximum render speed instead of capped frame rate

+++ esp32/src/effects/effect_processor.h
@@ -47,7 +47,7 @@ class EffectProcessor {
 	ProjectileEffect projectileEffect;
 	TextEffect textEffect;
 	ScrollTextEffect scrollTextEffect;
-	unsigned long lastFrameTime;
+	uint32_t lastFrameTime;  // Microseconds for high-precision timing
 
 	EffectEntry effectMap[9];
